requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.4.0=pypi_0
accelerate=1.12.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.13.3=pypi_0
aiosignal=1.4.0=pypi_0
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.12.1=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.4.0=pypi_0
bitsandbytes=0.49.2=pypi_0
bleach=6.3.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.12.2=h06a4308_0
certifi=2026.2.25=pypi_0
charset-normalizer=3.4.4=pypi_0
click=8.3.1=pypi_0
cloudpickle=3.1.2=pypi_0
cmake=4.2.1=pypi_0
cuda-bindings=12.9.4=pypi_0
cuda-pathfinder=1.4.0=pypi_0
datasets=4.6.1=pypi_0
dill=0.4.0=pypi_0
diskcache=5.6.3=pypi_0
exceptiongroup=1.3.1=pypi_0
expat=2.7.4=h7354ed3_0
fastapi=0.134.0=pypi_0
filelock=3.24.3=pypi_0
frozenlist=1.8.0=pypi_0
fsspec=2026.2.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.46=pypi_0
grpcio=1.78.0=pypi_0
h11=0.16.0=pypi_0
hf-xet=1.3.2=pypi_0
httpcore=1.0.9=pypi_0
httptools=0.7.1=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=1.5.0=pypi_0
idna=3.11=pypi_0
iniconfig=2.3.0=pypi_0
interegular=0.3.3=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.3=pypi_0
jsonschema=4.26.0=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
kaggle=1.7.4.5=pypi_0
lark=1.3.1=pypi_0
ld_impl_linux-64=2.44=h153f514_2
libexpat=2.7.4=h7354ed3_0
libffi=3.4.4=h6a678d5_1
libgcc=15.2.0=h69a1729_7
libgcc-ng=15.2.0=h166f726_7
libgomp=15.2.0=h4751f2c_7
libnsl=2.0.0=h5eee18b_0
libstdcxx=15.2.0=h39759b7_7
libstdcxx-ng=15.2.0=hc03a8fd_7
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
llvmlite=0.46.0=pypi_0
markdown=3.10.2=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.2=pypi_0
multidict=6.7.1=pypi_0
multiprocess=0.70.18=pypi_0
ncurses=6.5=h7934f7d_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=pypi_0
ninja=1.13.0=pypi_0
numba=0.64.0=pypi_0
numpy=2.2.6=pypi_0
nvidia-cublas-cu12=12.8.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.8.90=pypi_0
nvidia-cuda-nvrtc-cu12=12.8.93=pypi_0
nvidia-cuda-runtime-cu12=12.8.90=pypi_0
nvidia-cudnn-cu12=9.10.2.21=pypi_0
nvidia-cufft-cu12=11.3.3.83=pypi_0
nvidia-cufile-cu12=1.13.1.3=pypi_0
nvidia-curand-cu12=10.3.9.90=pypi_0
nvidia-cusolver-cu12=11.7.3.90=pypi_0
nvidia-cusparse-cu12=12.5.8.93=pypi_0
nvidia-cusparselt-cu12=0.7.1=pypi_0
nvidia-nccl-cu12=2.27.5=pypi_0
nvidia-nvjitlink-cu12=12.8.93=pypi_0
nvidia-nvshmem-cu12=3.4.5=pypi_0
nvidia-nvtx-cu12=12.8.90=pypi_0
omegaconf=2.3.0=pypi_0
openssl=3.5.5=h1b28b03_0
outlines=0.0.34=pypi_0
packaging=25.0=py310h06a4308_1
pandas=2.3.3=pypi_0
peft=0.18.1=pypi_0
pillow=12.1.1=pypi_0
pip=26.0.1=pyhc872135_0
platformdirs=4.9.2=pypi_0
pluggy=1.6.0=pypi_0
prometheus-client=0.24.1=pypi_0
propcache=0.4.1=pypi_0
protobuf=6.33.5=pypi_0
psutil=7.2.2=pypi_0
pthread-stubs=0.3=h0ce48e5_1
py-cpuinfo=9.0.0=pypi_0
pyarrow=23.0.1=pypi_0
pydantic=2.12.5=pypi_0
pydantic-core=2.41.5=pypi_0
pygments=2.19.2=pypi_0
pynvml=11.5.0=pypi_0
pytest=9.0.2=pypi_0
python=3.10.19=h6fa692b_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.2.1=pypi_0
python-slugify=8.0.4=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
ray=2.54.0=pypi_0
readline=8.3=hc2a1206_0
referencing=0.37.0=pypi_0
regex=2026.2.28=pypi_0
requests=2.32.5=pypi_0
rich=14.3.3=pypi_0
rpds-py=0.30.0=pypi_0
safetensors=0.7.0=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
sentencepiece=0.2.1=pypi_0
sentry-sdk=2.53.0=pypi_0
setuptools=80.10.2=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sqlite=3.51.1=he0a8d7e_0
starlette=0.52.1=pypi_0
sympy=1.14.0=pypi_0
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.6.0=pypi_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.22.2=pypi_0
tomli=2.4.0=pypi_0
torch=2.10.0=pypi_0
tqdm=4.67.3=pypi_0
transformers=5.2.0=pypi_0
triton=3.6.0=pypi_0
trl=0.29.0=pypi_0
typer=0.24.1=pypi_0
typer-slim=0.24.0=pypi_0
typing-extensions=4.15.0=pypi_0
typing-inspection=0.4.2=pypi_0
tzdata=2025.3=pypi_0
urllib3=2.6.3=pypi_0
uvicorn=0.41.0=pypi_0
uvloop=0.22.1=pypi_0
vllm=0.4.0=pypi_0
wandb=0.25.0=pypi_0
watchfiles=1.1.1=pypi_0
webencodings=0.5.1=pypi_0
websockets=16.0=pypi_0
werkzeug=3.1.6=pypi_0
wheel=0.46.3=py310h06a4308_0
xformers=0.0.23.post1=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xxhash=3.6.0=pypi_0
xz=5.6.4=h5eee18b_1
yarl=1.22.0=pypi_0
zlib=1.3.1=hb25bd0a_0
